Generate keypair at setup time and use that one
As pointed out in #1, reading the user's private key is a) unnecessary and
b) more work in the case of a passphrase-protected key. There's no real
reason why we need to use a pre-existing key, anyway. If we generate a priv+pub
keypair at setup time, we can then be confident that all preconditions are
met at connect time.